unshare automation_manager code
